\name{AffySNPPDInfoPkgSeed-class}
\docType{class}
\alias{AffySNPPDInfoPkgSeed-class}

\title{Class "AffySNPPDInfoPkgSeed"}
\description{
  This class represents Platform Design (PD) packages for Affymetrix
  mapping (SNP chip) arrays.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("AffySNPPDInfoPkgSeed", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{cdfFile}:}{Path to the CDF file for this
    platform.  Example: \code{"files/Mapping250K_Nsp.cdf"}}

    \item{\code{csvAnnoFile}:}{Path to the Affymetrix CSV annotation
    file.  Example: \code{"files/Mapping250K_Nsp_annot.csv"}}

    \item{\code{csvSeqFile}:}{Path to the Affymetrix probe sequence
      file.  Example: \code{Mapping250K_Nsp_probe_tab}}

    \item{splineParamFile}{Path to the spline parameters file used to
      compute the predicted accuracy of the the genotype calls.  Used
      internally in \code{.predictAccuracy}.}

    \item{crlmmInfoFile}{Path to is data file containing regions data
      used by the \code{crlmm} function.}

    \item{referenceDistFile}{Path to a reference distribution file used
      in the normalization step. This is the reference used in
      \code{snprma}.}

%% these are from PDInfoPkgSeed-class
    \item{\code{chipName}:}{Name of the chip or platform}
    \item{\code{manufacturer}:}{chip/platform manufacturer}
    \item{\code{url}:}{chip URL}
    \item{\code{genomebuild}:}{The genome build this platform is based
    upon.}
	\item{\code{organism}:}{ organism for chip. }
	\item{\code{species}:}{ species for chip. }  
%% these are from PkgSeed-class
    \item{\code{name}:}{The name of the package.}
    \item{\code{version}:}{A character vector giving the version
    number of the package.}
    \item{\code{license}:}{The license of the package}
    \item{\code{author}:}{Author of the package}
    \item{\code{email}:}{An email address to use in the Maintainer field}
    \item{\code{biocViews}:}{Character data for the biocViews field of
    the DESCRIPTION file}

  }
}
\section{Methods}{
	 \describe{
    \item{chipName}{\code{signature(object = "AffyExpressionPDInfoPkgSeed")}: ... }
    \item{getGeometry}{\code{signature(.Object = "AffyExpressionPDInfoPkgSeed")}: ... }
    \item{makePdInfoPackage}{\code{signature(object = "AffyExpressionPDInfoPkgSeed")}: ... }
	}
}

\section{Notes}{
  The files used for slots \code{splineParamFile}, \code{crlmmInfoFile},
  and \code{referenceDistFile} are generated by the Bioconductor project
  for each chip/platform and are hosted in our svn data repository at
  \url{https://hedgehog.fhcrc.org/bioc-data/trunk/annotation/parms_store}.
  When \code{makePdInfoPackage} is run, these files are simply copied to
  the \code{inst/extdata} directory of the generated package.
}

\author{Seth Falcon}
\examples{
showClass("AffySNPPDInfoPkgSeed")

cdfFile <- "Mapping250K_Nsp.cdf"
csvAnno <- "Mapping250K_Nsp_annot.csv"
csvSeq <- "Mapping250K_Nsp_probe_tab"

spline <- "pd.mapping250k.nsp.spline.params.rda"
refd <- "pd.mapping250k.nspRef.rda"
crlmmInf <- "pd.mapping250k.nspCrlmmInfo.rda"
\dontrun{
pkg <- new("AffySNPPDInfoPkgSeed", 
           version="0.1.5",
           author="A. U. Thor", email="au@thor.net",
           biocViews="AnnotationData",
           genomebuild="NCBI Build 35, May 2004",
           cdfFile=cdfFile, csvAnnoFile=csvAnno, csvSeqFile=csvSeq,
           splineParamFile=spline, crlmmInfoFile=crlmmInf,
           referenceDistFile=refd)

showMethods(classes=class(pkg))
}
}
\keyword{classes}
